# src/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"RepositoryDTO_unknown_": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {},
					"data": {}
				},
				"required": [
					"status",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DeleteModel": {
				"properties": {
					"ids": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					}
				},
				"required": [
					"ids"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RepositoryDTO_any_": {
				"properties": {
					"status": {
						"type": "number",
						"format": "double"
					},
					"message": {},
					"data": {}
				},
				"required": [
					"status",
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"TypeSort": {
				"enum": [
					0,
					1
				],
				"type": "number"
			},
...
			"UserModel": {
				"properties": {
					"email": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"phone"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"TheaterModel": {
				"properties": {
					"name": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"city": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"address",
					"city"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ShowtimeModel": {
				"properties": {
					"showDate": {
						"type": "string",
						"format": "date-time"
					},
					"endTime": {
						"type": "string"
					},
					"startTime": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"movieId": {
						"type": "number",
						"format": "double"
					},
					"screenId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"showDate",
					"endTime",
					"startTime",
					"price",
					"movieId",
					"screenId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"SeatModel": {
				"properties": {
					"row": {
						"type": "number",
						"format": "double"
					},
					"col": {
						"type": "number",
						"format": "double"
					},
					"seatNumber": {
						"type": "string"
					},
					"screenId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"row",
					"col",
					"seatNumber",
					"screenId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ScreenModel": {
				"properties": {
					"name": {
						"type": "string"
					},
					"seatCapacity": {
						"type": "number",
						"format": "double"
					},
					"theaterId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"seatCapacity",
					"theaterId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"MovieModel": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"duration": {
						"type": "number",
						"format": "double"
					},
					"releaseDate": {
						"type": "string",
						"format": "date-time"
					},
					"endDate": {
						"type": "string",
						"format": "date-time"
					},
					"trailerUrl": {
						"type": "string"
					},
					"posterUrl": {
						"type": "string"
					},
					"genreId": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					}
				},
				"required": [
					"title",
					"description",
					"duration",
					"releaseDate",
					"endDate",
					"trailerUrl",
					"posterUrl",
					"genreId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"GenreModel": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"description"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"JWT": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			},
			"tsoa_auth": {
				"type": "oauth2",
				"flows": {
					"implicit": {
						"authorizationUrl": "http://swagger.io/api/oauth/dialog",
						"scopes": {
							"write:pets": "modify things",
							"read:pets": "read things"
						}
					}
				}
			}
		}
	},
	"info": {
		"title": ""
	},
	"paths": {
		"/user": {
			"delete": {
				"operationId": "DeleteArray",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_unknown_"
								}
							}
						}
					}
				},
				"tags": [
					"User Controller"
				],
				"security": [
					{
						"JWT": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteModel"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetFilter",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_any_"
								}
							}
						}
					}
				},
				"tags": [
					"User Controller"
				],
				"security": [
					{
						"JWT": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "orderBy",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/TypeSort"
						}
					},
					{
						"in": "query",
						"name": "username",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_unknown_"
								}
							}
						}
					}
				},
				"tags": [
					"User Controller"
				],
				"security": [
					{
						"JWT": [
							"admin",
							"user"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserModel"
							}
						}
					}
				}
			}
		},
...
		"/Theater": {
			"get": {
				"operationId": "GetFilter",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_any_"
								}
							}
						}
					}
				},
				"description": "Lọc rạp chiếu phim",
				"tags": [
					"Theater Controller"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "orderBy",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/TypeSort"
						}
					},
					{
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "city",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "address",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Create",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_unknown_"
								}
							}
						}
					}
				},
				"tags": [
					"Theater Controller"
				],
				"security": [
					{
						"JWT": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TheaterModel"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteArray",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_unknown_"
								}
							}
						}
					}
				},
				"tags": [
					"Theater Controller"
				],
				"security": [
					{
						"JWT": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteModel"
							}
						}
					}
				}
			}
		},
...
		"/Showtime": {
			"get": {
				"operationId": "GetFilter",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_any_"
								}
							}
						}
					}
				},
				"description": "Lọc thể loại thời gian chiếu phim",
				"tags": [
					"Showtime Controller"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "orderBy",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/TypeSort"
						}
					},
					{
						"in": "query",
						"name": "showDate",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "movieId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "screenId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Create",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_unknown_"
								}
							}
						}
					}
				},
				"tags": [
					"Showtime Controller"
				],
				"security": [
					{
						"JWT": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ShowtimeModel"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteArray",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_unknown_"
								}
							}
						}
					}
				},
				"tags": [
					"Showtime Controller"
				],
				"security": [
					{
						"JWT": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteModel"
							}
						}
					}
				}
			}
		},
		"/Showtime/{id}": {
			"get": {
				"operationId": "GetOne",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_any_"
								}
							}
						}
					}
				},
				"tags": [
					"Showtime Controller"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_unknown_"
								}
							}
						}
					}
				},
				"tags": [
					"Showtime Controller"
				],
				"security": [
					{
						"JWT": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ShowtimeModel"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_unknown_"
								}
							}
						}
					}
				},
				"tags": [
					"Showtime Controller"
				],
				"security": [
					{
						"JWT": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/Seat": {
			"get": {
				"operationId": "GetFilter",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_any_"
								}
							}
						}
					}
				},
				"description": "Lọc ghế của rạp chiếu phim",
				"tags": [
					"Seat Controller"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "orderBy",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/TypeSort"
						}
					},
					{
						"in": "query",
						"name": "seatNumber",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "screenId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Create",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_unknown_"
								}
							}
						}
					}
				},
				"tags": [
					"Seat Controller"
				],
				"security": [
					{
						"JWT": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SeatModel"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteArray",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_unknown_"
								}
							}
						}
					}
				},
				"tags": [
					"Seat Controller"
				],
				"security": [
					{
						"JWT": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteModel"
							}
						}
					}
				}
			}
		},
...
		"/Screen": {
			"get": {
				"operationId": "GetFilter",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_any_"
								}
							}
						}
					}
				},
				"description": "Lọc phòng chiếu phim",
				"tags": [
					"Screen Controller"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "orderBy",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/TypeSort"
						}
					},
					{
						"in": "query",
						"name": "theaterId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Create",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_unknown_"
								}
							}
						}
					}
				},
				"tags": [
					"Screen Controller"
				],
				"security": [
					{
						"JWT": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ScreenModel"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteArray",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_unknown_"
								}
							}
						}
					}
				},
				"tags": [
					"Screen Controller"
				],
				"security": [
					{
						"JWT": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteModel"
							}
						}
					}
				}
			}
		},
...
		"/Review": {
			"get": {
				"operationId": "GetFilter",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_any_"
								}
							}
						}
					}
				},
				"description": "Lọc thể loại đồ ăn thêm tên,page,pagesize,với sắp xếp",
				"tags": [
					"Review Controller"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "orderBy",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/TypeSort"
						}
					},
					{
						"in": "query",
						"name": "userId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "rating",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "movieId",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Create",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_unknown_"
								}
							}
						}
					}
				},
				"tags": [
					"Review Controller"
				],
				"security": [
					{
						"JWT": [
							"admin",
							"user"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteArray",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_unknown_"
								}
							}
						}
					}
				},
				"tags": [
					"Review Controller"
				],
				"security": [
					{
						"JWT": [
							"admin",
							"user"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteModel"
							}
						}
					}
				}
			}
		},
...
		"/Movie": {
			"get": {
				"operationId": "GetFilter",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_any_"
								}
							}
						}
					}
				},
				"description": "Lọc phim",
				"tags": [
					"Movie Controller"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "orderBy",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/TypeSort"
						}
					},
					{
						"in": "query",
						"name": "title",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "genreId",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"in": "query",
						"name": "statusMovieEnum",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Create",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_unknown_"
								}
							}
						}
					}
				},
				"tags": [
					"Movie Controller"
				],
				"security": [
					{
						"JWT": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MovieModel"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteArray",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_unknown_"
								}
							}
						}
					}
				},
				"tags": [
					"Movie Controller"
				],
				"security": [
					{
						"JWT": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteModel"
							}
						}
					}
				}
			}
		},
...
		"/Genre": {
			"get": {
				"operationId": "GetFilter",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_any_"
								}
							}
						}
					}
				},
				"description": "Lọc thể loại phim theo tên",
				"tags": [
					"Genre Controller"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "orderBy",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "sort",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/TypeSort"
						}
					},
					{
						"in": "query",
						"name": "name",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Create",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_unknown_"
								}
							}
						}
					}
				},
				"tags": [
					"Genre Controller"
				],
				"security": [
					{
						"JWT": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GenreModel"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteArray",
				"responses": {
					"204": {
						"description": "No content",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/RepositoryDTO_unknown_"
								}
							}
						}
					}
				},
				"tags": [
					"Genre Controller"
				],
				"security": [
					{
						"JWT": [
							"admin"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeleteModel"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/api"
		}
	]
}